requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
fastapi==0.103.2
filelock==3.12.4
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
identify==2.5.30
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
virtualenv==20.24.5
python-dotenv
python-multipart
aiofiles
pillow
pytesseract
